Added virus checking from the supplied filename in the scan bucket
We want to virus check files which have been uploaded to the scan bucket
in s3. When the task is received the file is downloaded and we use the
clamd python library to talk to the clam av on the docker image over a
unix socket. If the check is success it posts to the scan-passed queue
otherwise to the scan-failed queue for processing.

* Added virus checking client which uses clamd to virus the file
* Updated run script to start clamd and fleshclam (the database updater)
* Updated conftest.py to handle the context and pytest.ini to load in
static config for testing
* Added tests for the anti-virus client