Add an API endpoint to scan files synchronously
For document download we need to perform the virus scan before we
encrypt and store the document in S3, so we need an API endpoint
in addition to the worker task.

This adds a `/scan` endpoint that accepts a document upload, scans
it and returns `true` if the document is virus-free or `false` if
it's not.