Add a PaaS application for deploying antivirus as an API
Adds a separate application to the PaaS manifest that would run the
gunicorn server instead of the celery worker. This is similar to the
approach used by notifications-api and allwos us to scale API and
workers independently from each other.

This also modifies the deployment make steps so that they could be
used to deploy workers or the API by setting NOTIFY_APP_NAME.