Fix empty variable error in PaaS manifest generation
generate-manifest.py used pyyaml to generate the manifest yaml file,
which means it was aware of the value types and would escape/quote
them when necessary. So e.g. if a variable value is an empty string
the script would insert `''` into the output.

Since Jinja is not aware of the yaml semantics of the output file
it inserts values as is. So an empty string produces no output,
making it equivalent to a `null` value. This is invalid for an
environment value, so CF CLI raises an error when trying to use
the manifest.

To avoid this, we explicitly quote all environment variable values.