Split workers and API manifests into separate templates
Older versions of CF CLI do not support additional arguments when
deploying one application from a multi-app manifest, which we need
in order to deploy a docker container.

This is fixed in CF CLI >6.34, but we can't upgrade the CLI package
on our CI instance since the new version of CF CLI breaks process
substitution when used with deprecated manifest features (like top
level keys), which are used in api, admin and template preview apps.

To avoid the issue for now we split the manifest template into 2
single-app manifests: one for the celery app and one for the API.