Updated the Dockerfile to be a multi-stage build and therefore take away the need for having two Dockerfiles one for test and one for production.
First the base image is produced and then two separate build one for
test and one for production both using the parent image as the base for
the specific implementation.

This allows the production image to be a lot smaller and reduce the risk
of sensitive information being copied as only the white listed list
files are copied to the image. It would allow the for different packages
etc being install per image.

* Updated Dockerfile to use mutli-stage build
* Remove the old Dockerfile_for_test file
* Updated the makefile to use the build stage build and not the test
Dockerfile