Monitor clamav processes and restart them on failure
`start_application` used to collect pids for freshclam, clamd and
celery workers into one variable that was used to check that the
app is still alive.

The variable is used as an argument to `kill -0` in order to check
that the processes are still alive. Unfortunately, the version of
`kill` we're running on PaaS seems to return a 0 exit code if at
least one of the processes is alive (it only returns 1 if all of
the argument PIDs don't refer to an existing process). This means
that if clamd or freshclam crash (or are killed by OOM killer) the
startup script ignores them and keeps the app running, preventing
the instance from being replaced.

We could capture the variables separately and stop the startup script
loop if any of the processes disappear, but replacing an instance
would take longer than restarting the crashed process, so instead we
split the startup functions so that we can monitor and restart clamd
and freshclam separately.