Don't run clamd as a background process
The docs say that clamd should not be started as a background process -
instead you should run `clamd` which will load the database and then
daemonize itself.

This change means that we can't rely on getting clamd's PID to check if
it is running, because the PID will change once clamd daemonizes itself.
To check if it's running we now ping clamd instead, and we also add this
as a check to the `/_status` endpoint.