Bump clamav version to fix build
Our build was passing on jenkins because it was
using it's cache. However when trying to run this on
concourse it did not have the old version of clamav in
its cache and failed to build as the version is no longer
available through apt-get.

Two versions are available through apt-get at the moment:
clamav-daemon/oldoldstable 0.101.4+dfsg-0+deb8u2 amd64
clamav-daemon/oldoldstable 0.100.0+dfsg-0+deb8u1 amd64

I've bumped us to the latest.