update dockerfile - bump debian and add pycurl dependencies
we flag celery[sqs] in reqs.txt, as it has additional dependencies to
install (pycurl). Pycurl requires libssl-dev and libcurl4-openssl-dev
debian packages, which aren't available in jessie. Jessie is
"oldoldstable" now, it's probably a good time to update the dockerfile
to run on regular stable (which is currently buster).

CLAMAV_VERSION defines what package to get - because we append a * on to
it we don't need to specify all the flags - that last flag in particular
is pinning clamav to the debian 8 (jessie) version. We're just
interested in the 0.101.4 bit.