Ignore Celery logs by redirecting to /dev/null
We want to ignore the logs that come from Celery itself since we
already have logs that cover the tasks - the Celery logs provide
duplicate data. This copies the approach we use in our delivery apps
and redirects the Celery logs (which go to stderr by default) to
`/dev/null`.