Switch to a script for running Docker commands
Previously we were limited by the fixed set of rules provided by
the Makefile. While the 'bash-with-docker' rule could be used to
subsequently run arbitrary commands, that was more effort. This
moves and simplifies the 'docker' function from a Makefile function
to a script, so that it can be run with arbitrary commands.

Note that the previous 'WERKZEUG_DEBUG_PIN' setting is inconsistent
with our other projects. We can always re-add it if we need it.