Squash '_' make rules into Docker ones
This simplifies the make rules to run app commands by removing
multiple levels of rules. Note that we need to keep the scripts,
as starting this app involves running a separate 'clamd' command.

This also simplifies the command to run Celery, so that we get
more logging output in development, as we do in other apps [1].

[1]: https://github.com/alphagov/notifications-template-preview/commit/feae0437828b61a98054d3419ca0dbd274fa69e9#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52R67