error if we get rate limited when updating clamav
Because concourse doesn't cache docker images, we always have to run
freshclam to update from scratch every time we build a docker image.
Normally this happens once a day, however, we sometimes build docker
images more (if a later step of the docker file fails, or if we merge a
change to antivirus). 

If it's rate limited, freshclam blocks updates for 24 hours, but it
doesn't return an error code. Our docker build carries on but with no
virus definitions, and when the app is deployed the functional tests
will fail (correctly). This blocks all our pipelines until the rate
limit clears and we can build a working docker image. We'd much rather
just have to stay on an old image for a bit longer if we're
rate-limited. To do so, we can detect this rate limit and error out of
the docker build, meaning concourse will still use the old image for
other dependent jobs.

Instead of checking error codes, instead we need to parse the log lines
to look for the 429 error message. To do so we redirect output from
clamav to stdout so grep can see it, but then tee it back to stderr so
we still get viewable logs in our build output.

This behaviour from freshclam is included in their test cases (including
the error code being 0) so it feels unlikely that this log line will
change much soon. [^1]

--- 

[^1]: https://github.com/Cisco-Talos/clamav/blob/30561caff531dd7d0c16aba214399baf3fceab40/unit_tests/freshclam_test.py#L277-L282