Fix build passing when freshclam hits a 429
Using a piped command to fix this doesn't seem to work - we've had
several passing builds where freshclam has 429ed [^1].

This technique does work [^2].

The piped command also had a problem that it would "succeed" if the
"freshclam --stdout" command failed, due to the "!" at the start e.g.

      #5 [2/2] RUN ! eho "hi error" |    tee /dev/stderr |   grep -q error
      #5 sha256:db7f997518950e9fff1cf982bfff42a1c13c91b0c85da0ac5c031b6095c88b76
      #5 0.172 /bin/bash: line 1: eho: command not found
      #5 DONE 0.2s

Using this technique fixes that too.

[^1]: https://concourse.notify.tools/teams/notify/pipelines/apps/jobs/build-antivirus-master/builds/155
[^2]: https://concourse.notify.tools/teams/notify/pipelines/apps/jobs/build-antivirus-master/builds/158