Download Clamav signature database from s3 using PrivateMirror
We are consistently hitting Clamav's signature database rate limit
as we download a fresh copy with each build, often multiple times
a day.

The recommended solution is to maintain a database mirror.

Using the PrivateMirror option, we can instead bypass the DNS signature
checks and simiply use the if-motified-since header to detect
database updates. This allows us to use S3 as a suitable repository.

We can then periodically update the database signature files stored
in S3 via an external process.

This will allow us to build this project multiple times a day without
exceeding the rate limit.

PrivateMirror details:

```
This option allows you to easily point freshclam to private mirrors. If PrivateMirror is set, freshclam does not attempt to use DNS to determine whether its databases are out-of-date, instead it will use the If-Modified-Since request or directly check the headers of the remote database files. For each database, freshclam first attempts to download the CLD file. If that fails, it tries to download the CVD file. This option overrides DatabaseMirror, DNSDatabaseInfo and ScriptedUpdates. It can be used multiple times to provide fall-back mirrors.
```