Make sure all existing packages are up-to-date
This is part of the work to setup image scanning to report any
security vulnerabilities, which we then need to fix. Currently
I'm seeing 6 issues reported for low-level OS packages. Adding
each package to the install list to upgrade it doesn't scale.

The base image should eventually update with newer versions, but
this seems to be slower than the rate of vulnerabilities, which
we want to fix immediately if we can. One way to apply the fixes
is to upgrade all base image packages to their latest versions.

Ideally we would only apply security updates, but I can't find a
simple way of doing this. Since we were letting the versions drift
anyway, the additional risk should be minimal. There's a fairly
compelling argument for the upgrade approach online [^1].

[^1]: https://pythonspeed.com/articles/security-updates-in-docker/