Add entrypoint command to API manifest
In commit 3f19dc7a05e5fea2e7d31a6afb9024937dccbe36 we changed from
specifying the command for api/celery in the manifest to using an
entrypoint script. But we forgot to include that entrypoint script as
the command for the manifest.

This meant that we were still running the apps with the old commands.
When we deleted the CF apps and had to recreate them, they didn't have
a specific command to run and were just defaulting to python3.