Set `setuid` bit on `clamd`
chown `clamd` to the `clamav` user and `clamav` group, then set the
`setuid` bit on it. This will allow the `clamav` user (or anyone in the
`clamav` group) to run the `clamd` executable, and when they do so, it
will run as the `clamav` user.

This will let us have the `notify` start the clam daemon, but then have
it running as the `clamav` user, which means we don't need to wrangle
any of the clamav dir permissions.