Bump instance count to 3
We had an outage recently due to PaaS rolling our antivirus instances
without waiting for a replacement to come up.

This is not a fix, but would give paas more time to perform the roll
successfully before potentially causing an outage for us.

The exact cause of the incident is unknown but may be down to slow
docker image pull times from github during the app rolling.