Use uv to install packages in Docker
In order to do this we need to install uv inside the Docker container.
We don't want it being installed in the venv that we use for production.
So we need to do this as a separate step.

Part of the reason uv is faster is because it uses a cache of packages.
By default it uses `$HOME/.cache/uv` as the location for this cache
(even if you use --no-cache, it makes a temp folder in there)[^1].

Our docker image doesn't have permission to write there, so we override
the default and set the cache to be in /tmp/ instead

[^1] https://docs.astral.sh/uv/concepts/cache/#cache-directory